Exclude some classes from coverage
