+++ src/main/java/eu/dissco/core/digitalspecimenprocessor/configuration/ApplicationConfiguration.java
@@ -3,10 +3,6 @@ package eu.dissco.core.digitalspecimenprocessor.configuration;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import eu.dissco.core.digitalspecimenprocessor.component.DateDeserializer;
-import eu.dissco.core.digitalspecimenprocessor.component.DateSerializer;
-import eu.dissco.core.digitalspecimenprocessor.component.InstantDeserializer;
-import eu.dissco.core.digitalspecimenprocessor.component.InstantSerializer;
 import java.time.Instant;
 import java.util.Date;
 import java.util.Random;
