+++ src/main/java/eu/dissco/core/digitalspecimenprocessor/configuration/DateDeserializer.java
@@ -1,4 +1,4 @@
-package eu.dissco.core.digitalspecimenprocessor.component;
+package eu.dissco.core.digitalspecimenprocessor.configuration;
 
 import static eu.dissco.core.digitalspecimenprocessor.configuration.ApplicationConfiguration.DATE_STRING;
 
@@ -9,16 +9,19 @@ import java.io.IOException;
 import java.time.Instant;
 import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
+import java.util.Date;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-public class InstantDeserializer extends JsonDeserializer<Instant> {
+public class DateDeserializer extends JsonDeserializer<Date> {
+
   private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern(DATE_STRING).withZone(
       ZoneOffset.UTC);
+
   @Override
-  public Instant deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) {
+  public Date deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) {
     try {
-      return Instant.from(formatter.parse(jsonParser.getText()));
+      return Date.from(Instant.from(formatter.parse(jsonParser.getText())));
     } catch (IOException e) {
       log.error("An error has occurred deserializing a date. More information: {}", e.getMessage());
       return null;
